Refactor parameter inference (#358)
* Initial commit with CHANGELOG.rst

* Remove methods.parameter_inference.py

* Add inference-folder containing methods.

* Make methods public from new folder.

* Fix tests given the new folder-structure.

* Fix linting.

* Fix file description comments.
