+++ elfi/__init__.py
@@ -13,7 +13,9 @@ import elfi.model.tools as tools
 from elfi.client import get_client, set_client
 from elfi.methods.diagnostics import TwoStageSelection
 from elfi.methods.model_selection import *
-from elfi.methods.parameter_inference import *
+from elfi.methods.inference.bolfi import *
+from elfi.methods.inference.romc import *
+from elfi.methods.inference.samplers import *
 from elfi.methods.post_processing import adjust_posterior
 from elfi.model.elfi_model import *
 from elfi.model.extensions import ScipyLikeDistribution as Distribution
